Added genetic alghoritm, LCG and MTcrack

+++ PseudoRandomDecoder/CasinoResponse.cs
@@ -4,6 +4,6 @@
     {
         public string message { get; set; }
         public Account account { get; set; }
-        public int realNumber { get; set; }
+        public long realNumber { get; set; }
     }
 }
